Address feedback from MR

+++ models/video_utils.py
@@ -25,7 +25,7 @@ def get_numpy(x: Tensor) -> np.ndarray:
     return x.squeeze().cpu().numpy()
 
 def non_zero_mean(x: Tensor) -> float:
-    return sum(x) / len(x) if len(x) > 0 else -1
+    return float(sum(x) / len(x) if len(x) > 0 else -1)
 
 def compute_psnr(prediction: Tensor, target: Tensor) -> float:
     """
